collapse directives to 3 basic cases
Prior to this commit the directive were extremely repetitive without a
great deal of value and, in some cases, hiding the relationship between
the Directive type and the related Parser function. This commit distills
the directives into 3 basic cases:
* Directive type == Parser function
* Directive type != Parser function
* Directive type != Parser function, add zulu flag

Although none currently exist, if we encounter a case that does not fit
any of these 3 scenarios, then we can add it as a special case.