add support for partial-hour offset timezones
Prior to this commit Timex would produce a blantantly incorrect value
for parsing a timestamp with a partial-hour offset timezone, such as
that of Nepal (+05:45). The resulting timestamp would have the correct
wall time, but a truncated timezone (+05:00), which would result in
incorrect timestamps in UTC. This commit adds support for partial-hour
offset timezones by calling them out as special cases when attempting to
execute `Timex.Timezone.name_of`, which eventually gets invoked by
`Timex.parse`.