Add typespecs for Timex.Protocol and adjust irregularities
Adding the typespecs to the protocol will enable dialyzer
to verify the typespecs of the implementations as well as
the calls from the main Timex module.

The typespecs in the Timex module was updated to use
Types.valid_datetime instead of the more verbose multi type
version (also Types.calendar_types does not exist).

Error tuple returns was added to those missing it in the
Timex module.

The protocol implementations type specifications is
narrowed down to match those of the protocol/caller module.

The typespecs for to_julian was changed to float, which is
what is returned.

Also fixes a few bugs:

* from_iso_day for Date now returns a Date struct instead of a map
* beginning_of_quarter for Date now works instead of crashes
* beginning_of_quarter now returns the correct time for erlang datetimes
* end_of_week now returns a erlang datetime instead of a date for erlang datetimes