Remove ok! and try! macros
The ok! macro was only used with Timex.to_gregorian_microseconds which
does never return an ok-tuple, so we can just test against the error
tuple or assume ok return value.

Also moves to_gregorian_microseconds-call and error checking to diff
module. This makes the code easier to read and fixes a few cases where
the error tuple from that call was not handled/returned.