Fix issue #349: crash on `Timex.local` on input with ambiguous wall time
Previously, the timezone checked for ambiguity on the input date before
converting it to the actual timezone, so the local function could get
an AmbiguousTimezoneInfo for dates that aren't actually ambiguous in
the target timezone.

As far as I can see, `Timex.local/1` is just syntactic sugar for
`Timex.Timezone.convert(date, :local)`, which is the solution I've used.