Fix #356 - parsing of milliseconds
The use of `ms |> String.trim("0")` could result in an empty string when
the input was "000", which in turn made `String.to_integer` complain.

Since `trim` removes zeros from both ends of the String, milliseconds
ending in `0` or `00` were incorrectly parsed as well.

This commit changes `String.trim` to `String.trim_leading` and adds
a safety-clause to protect against the "000" input.