fix: issue where non-olson timezones could not be operated on via standard apis
The root cause here was that Timex was delegating responsibility for
timezones to tzdata, but supporting non-standard timezones in some cases
for compatibility with common formats. When passing DateTimes with such
timezones to standard library APIs like `DateTime.add`, the tzdata
timezone database would return :timezone_not_found because it doesn't
understand these zones, which in turn caused those operations to fail.
The fix here is to have Timex itself act as a timezone database, and
delegate to tzdata under the covers, kicking in as a fallback when
tzdata doesn't understand the zone it is given.

In addition, I've added proper support for POSIX-TZ timezone names, so
that we are correctly computing periods for these timezones to feed back
for the `Calendar.TimeZoneDatabase` behaviour.