Add latvian translations
saw that those were missing, and was keen on using them for stuff, so I
figured I'd be nice and make them a thing.

Some things of note.
1) Latvians don't really use the 1000 separator, it's not really a thing
over here
2) You might notice that the abbreviations are somewhat unusual (e.g.
march and may aren't abbreviated) that's because apparently according to
the official language center documents those months _aren't_ abbreviated
(see
http://www.vvc.gov.lv/advantagecms/export/docs/noderigi/saisinajumi/saisinajumi-1.pdf)
additionally days of the week are generally abbreviated to a single
letter, and second one is added only to the second day of that week that
begins with said letter (e.g. S - Sestdiena (Saturday)  and Sv - Svētdiena (Sunday)